Renaming separate model folders to just models so all models are found in the same place.
