+++ fileserve/blueprints/fileserve.py
@@ -3,7 +3,7 @@
 import os
 from flask import Blueprint, render_template, send_file, session, redirect, url_for
 from .. import conf
-from .. import filemodel
+from ..models.filelist import FileList
 
 api = Blueprint('fileserve', __name__, url_prefix='/')
 
@@ -13,7 +13,7 @@ def index():
     if 'username' not in session:
         return redirect(url_for('auth_controller.login'))
 
-    filemodel_list = filemodel.filelist.FileList('/')
+    filemodel_list = FileList('/')
     return render_template('index.html', files=filemodel_list)
 
 @api.route('/<path:path>')
@@ -23,6 +23,6 @@ def file_path(path):
         return redirect(url_for('auth_controller.login'))
     fileserve_source = conf.BaseConfig.FILES_BASE_PATH
     if os.path.isdir(fileserve_source + path):
-        filemodel_list = filemodel.filelist.FileList(path)
+        filemodel_list = FileList(path)
         return render_template('index.html', files=filemodel_list)
     return send_file(fileserve_source + path)
