--- fileserve/filemodel/__init__.py
@@ -1,3 +1,0 @@
-""" Filemodel represents a given file and all attributes """
-from . import filemodel
-from . import filelist
