--- fileserve/user_model/__init__.py
@@ -1,3 +1,0 @@
-""" init user_model in project """
-
-from . import user_model
